Introduced simple logging, added swapchain
Introduced logging via the log crate and the env_logger. Plans are to
replace env_logger with fern in the future. For now env_logger is fine
and gets the job done.
Added the swapchain to the main struct.
Fiddeling with CommandBuffers. I'm not sure how to handle those yet, as
they behave very differently in vulkano than in the normal C API.