Drawing a triangle!
I just wanted to render something and implemented the missing elements
like the pipeline. It is a static pipeline and nothing fancy happens.
A DynamicState is used, that way the GraphicsPipeline does not need to be
rebuilt, an expensive step. The now implemented Vertex struct is just
temporary to enable 2D rendering. Logging isn't implemented in the new
functions.